[GC-33] Fix the failing tests
* To fix the failing tests, we need to rewrite/refactor lots of
production code. The existing design of the production code is BAD. It
is not testable.

* When the application starts, a ContactManager object will be created
and in its constructor, we try looking for the contacts.bin file. The
file path and name is hard-coded and upon testing we don't have any
control specifying different path. So as part of testing, we end-up
creating contacts.bin file in the production code.

* Refactored the code, so that the contacts.bin file path can be passed
as parameter upon ContactManager creation. This way, the testing code
can send a different file name and or location for the contacts file
without affecting the production code contacts file.

* All the tests pass now, and implementing further tests should be easy.