[GC-38] Label for phone numbers
* Phone numbers may have an optional label.
* The label could be any valid string or null or empty string.
* On the surface, it seems to be a key-value pair and hashmap might be
the good fit, but it is NOT. The label may be null and so it cannot be
used as a key. Phone number can be used as key, but it gives a wrong
impression that, given a key (phone number) the value (label) can be
retrieved. But it is quite the opposite. So we used the Pair class from
Apache Commons lang library.